requirements.txt
streamlit
pdfplumber
nltk
scikit-learn
openai
python-dotenv
